pyproject.toml
[project]
name = "heatwaved-cli"
version = "0.1.0"
description = "A CLI tool for Oracle MySQL HeatWave POC demonstrations"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "cryptography>=45.0.4",
    "mysql-connector-python>=9.3.0",
    "oci>=2.154.3",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "typer>=0.16.0",
]

[project.scripts]
heatwaved = "heatwaved.main:app"

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]